refactor(#834): split the site into small files; fix three untrue claims
styles.css (15k) becomes 18 files under css/, one per page area, linked in
cascade order — responsive.css last so the cross-section phone rules still win.
app.js becomes ES modules under js/, state.js holding the live release/gear
values so i18n.js and gear.js share them without a cycle. Each i18n dictionary
splits into four files under i18n/<lang>/, merged at apply time.

Copy fixes, all of them things the site claimed that aren't true:
- "no hardware to buy / the laptop you already own is the rig" — you still
  need the audio interface to plug the guitar into. Says so now.
- "tell Claude you want the sound off a record" — meaningless. It's a song or
  an artist name, which is what the card below it already said.
- tone.lede said OpenRig captures the amp, as if capture were the mechanism.
  There are native amps written in Rust too; both are named now.
- captures.h2 invented a number of files. Dropped.

English markup fallbacks re-synced with the dictionaries so nobody reads a
retired sentence before the JSON lands.
