+++ site/css/band.css
@@ -1,0 +1,6 @@
+.band{position:relative;min-height:min(82svh,760px);display:flex;align-items:center;overflow:hidden}
+.band-media{position:absolute;inset:0}
+.band-media img{width:100%;height:100%;object-fit:cover;filter:saturate(.8) contrast(1.05)}
+.band-media::after{content:'';position:absolute;inset:0;background:linear-gradient(to right,rgba(5,7,11,.95) 0%,rgba(5,7,11,.8) 38%,rgba(5,7,11,.15) 78%)}
+.band .wrap{position:relative;z-index:2}
+.band-copy{max-width:620px}
