+++ site/css/buttons.css
@@ -1,0 +1,8 @@
+.btn{display:inline-flex;align-items:center;justify-content:center;gap:9px;font-weight:700;font-size:14px;letter-spacing:.06em;text-transform:uppercase;padding:13px 24px;border-radius:2px;transition:transform .25s var(--ease),background .25s,box-shadow .25s;white-space:nowrap}
+.btn-primary{background:var(--amber);color:#180c00;box-shadow:0 8px 30px -10px rgba(255,158,61,.7)}
+.btn-primary:hover{background:var(--amber-glow);transform:translateY(-2px)}
+.btn-ghost{border:1px solid var(--line2);color:var(--white)}
+.btn-ghost:hover{border-color:var(--white);transform:translateY(-2px)}
+.btn-support{border:2px solid var(--amber);color:var(--amber)}
+.btn-support:hover{background:rgba(255,158,61,.15);transform:translateY(-2px)}
+.btn-lg{padding:18px 34px;font-size:15px}
