+++ site/css/features.css
@@ -1,0 +1,10 @@
+.feat-grid{display:grid;grid-template-columns:repeat(3,1fr);gap:22px;margin-top:52px}
+.feat-card{border:1px solid var(--line);border-radius:5px;overflow:hidden;background:var(--panel);box-shadow:0 30px 70px -45px #000;transition:border-color .3s,transform .3s var(--ease)}
+.feat-card:hover{border-color:var(--line2);transform:translateY(-4px)}
+.feat-card .feat-shot{background:var(--ink2);display:flex;align-items:center;justify-content:center;height:220px;overflow:hidden}
+.feat-card .feat-shot img{width:100%;height:100%;object-fit:cover;object-position:top}
+.feat-card .feat-body{padding:20px 22px 24px}
+.feat-card h3{color:#fff;font-size:clamp(20px,1.8vw,26px);margin-bottom:8px}
+.feat-card p{font-size:14.5px;color:var(--mute)}
+@media(max-width:900px){.feat-grid{grid-template-columns:1fr 1fr}}
+@media(max-width:600px){.feat-grid{grid-template-columns:1fr}}
