+++ site/css/hero.css
@@ -1,0 +1,31 @@
+.hero{position:relative;min-height:100svh;display:flex;align-items:flex-end;overflow:hidden}
+.hero-media{position:absolute;inset:0}
+.hero-media img{width:100%;height:100%;object-fit:cover;object-position:58% 42%;filter:saturate(.9) contrast(1.05) brightness(1.12)}
+.hero-media::after{content:'';position:absolute;inset:0;
+  background:
+   linear-gradient(to top,rgba(5,7,11,.98) 0%,rgba(5,7,11,.78) 26%,rgba(5,7,11,.22) 58%,rgba(5,7,11,.55) 100%),
+   linear-gradient(to right,rgba(5,7,11,.82) 0%,rgba(5,7,11,.32) 45%,rgba(5,7,11,0) 75%);}
+.hero .wrap{position:relative;z-index:2;padding-bottom:clamp(56px,7vh,96px);padding-top:140px}
+.hero h1{color:#fff;margin:16px 0 0;text-shadow:0 4px 60px rgba(0,0,0,.6)}
+.hero h1 em{font-style:normal;color:var(--amber)}
+.hero .lede{margin-top:22px}
+.hero-cta{display:flex;flex-wrap:wrap;gap:14px;align-items:center;margin-top:34px}
+.hero-note{font-size:13px;color:var(--mute);letter-spacing:.04em}
+.scroll-hint{position:absolute;right:max(4vw,24px);bottom:38px;z-index:2;display:flex;align-items:center;gap:12px;font-size:11px;letter-spacing:.24em;text-transform:uppercase;color:var(--mute);writing-mode:vertical-rl}
+.scroll-hint i{display:block;width:1px;height:64px;background:linear-gradient(var(--amber),transparent)}
+
+.hero-product{position:absolute;top:104px;right:max(4vw,24px);z-index:2;width:min(36vw,440px)}
+.hero-product img{width:100%;display:block;border:1px solid var(--line2);border-radius:8px;background:var(--panel);box-shadow:0 50px 100px -35px rgba(0,0,0,.85)}
+/* The compact view sits under the full window, overlapping it — one glance shows both. */
+.hero-product img.hero-product-2{width:68%;margin:-8% 0 0 auto;transform:translateX(8%)}
+@media(max-width:1000px){.hero-product{display:none}}
+
+/* sponsor call — sits under the product shots, loud enough to be seen first */
+.hero-support{display:block;margin-top:18px;padding:18px 22px;border:2px solid var(--amber);border-radius:6px;background:rgba(255,158,61,.10);backdrop-filter:blur(8px);text-align:center;transition:background .25s,transform .25s var(--ease)}
+.hero-support:hover{background:rgba(255,158,61,.2);transform:translateY(-2px)}
+.hero-support span{display:block;font-family:var(--display);font-size:clamp(24px,2.1vw,32px);letter-spacing:.04em;text-transform:uppercase;color:var(--amber)}
+.hero-support i{display:block;font-style:normal;font-size:12.5px;letter-spacing:.05em;color:var(--soft);margin-top:5px}
+/* Below 1001px the card is hidden, so the CTA row carries the ask instead. */
+@media(min-width:1001px){.btn-support{display:none}
+  /* the sponsor card occupies that corner — the scroll hint would sit on top of it */
+  .scroll-hint{display:none}}
