+++ site/css/roadmap.css
@@ -1,0 +1,11 @@
+.road{display:grid;grid-template-columns:repeat(3,1fr);gap:clamp(24px,3vw,48px);margin-top:56px}
+.road-h{font-family:var(--display);font-size:20px;letter-spacing:.06em;text-transform:uppercase;padding-bottom:12px;border-bottom:1px solid var(--line);display:flex;align-items:center;gap:10px}
+.road-h::before{content:'';width:9px;height:9px;border-radius:50%;flex:none}
+.road-h.done::before{background:#4ade80}
+.road-h.next::before{background:var(--amber)}
+.road-h.later::before{background:#5d6c81}
+.road-col ul{list-style:none;margin:18px 0 0;padding:0}
+.road-col li{position:relative;padding:0 0 0 18px;margin-bottom:11px;font-size:15px;line-height:1.5;color:var(--mute)}
+.road-col li::before{content:'';position:absolute;left:0;top:.65em;width:6px;height:1px;background:var(--line2)}
+.road-col.rv .road-h.done+ul li{color:#c6d0dc}
+@media(max-width:820px){.road{grid-template-columns:1fr}}
