+++ site/css/stats.css
@@ -1,0 +1,10 @@
+.stats{display:grid;grid-template-columns:repeat(6,1fr);gap:1px;background:var(--line);border:1px solid var(--line)}
+.stat{background:var(--ink2);padding:34px 26px}
+.stat b{display:block;font-family:var(--display);font-size:clamp(40px,4.4vw,62px);color:#fff;line-height:1}
+.stat i{display:block;font-style:normal;font-size:12px;letter-spacing:.2em;text-transform:uppercase;color:var(--mute);margin-top:10px}
+
+.gearstats{display:grid;grid-template-columns:repeat(6,1fr);gap:1px;background:var(--line);border:1px solid var(--line);margin-top:44px}
+.gearstats div{background:var(--ink2);padding:22px 16px}
+.gearstats b{display:block;font-family:var(--display);font-size:clamp(26px,2.4vw,36px);color:#fff;line-height:1}
+.gearstats i{display:block;font-style:normal;font-size:11px;letter-spacing:.14em;text-transform:uppercase;color:var(--mute);margin-top:8px}
+@media(max-width:760px){.gearstats{grid-template-columns:repeat(3,1fr)}}
