+++ site/css/type.css
@@ -1,0 +1,9 @@
+h1,h2,h3,.dsp{font-family:var(--display);font-weight:400;letter-spacing:.012em;line-height:.92;text-transform:uppercase}
+h1{font-size:clamp(58px,8.6vw,148px)}
+h2{font-size:clamp(42px,5.6vw,92px)}
+h3{font-size:clamp(26px,2.4vw,36px)}
+.eyebrow{font-family:var(--body);font-size:12px;font-weight:700;letter-spacing:.26em;text-transform:uppercase;color:var(--amber)}
+.lede{font-size:clamp(17px,1.35vw,21px);line-height:1.55;color:var(--soft);max-width:56ch}
+.sub{color:var(--mute)}
+.code{font-family:ui-monospace,SFMono-Regular,Menlo,monospace;font-size:13.5px;color:var(--soft);background:var(--ink2);border:1px solid var(--line);border-radius:5px;padding:18px 20px;line-height:1.9;overflow-x:auto}
+.code .c{color:#5f6d82}
