+++ site/i18n/en/end.json
@@ -1,0 +1,59 @@
+{
+  "story.eyebrow": "How it started",
+  "story.h2": "This started because three<br>multi-effects units cost a fortune.",
+  "story.p1": "My son started guitar lessons. His teacher had one sound, we had another one at home, and what he learned in the lesson didn't happen again afterwards. Fixing that by buying everyone the same unit cost more than the instrument.",
+  "story.p2": "I've been a developer for 26 years. At the time I was investigating, with method, how far you can go programming with AI — and a study like that is only worth anything if the problem is hard. Live audio is hard: the code runs a few hundred times a second and, when it gets it wrong, the sound fails right there, in front of everyone. There's no hiding it.",
+  "story.p3": "It worked out better than I expected. OpenRig today is a real-time audio platform written in Rust: every input is an isolated runtime, so two guitars on the same interface really are two independent rigs. Amp, cab and pedal mixing native DSP, neural captures and IRs in the same chain. Presets in plain text you can open and edit.",
+  "story.p4": "And the original idea grew with it: instead of one tone for me and my son, one computer on stage processing the whole band, each musician working on their own sound. Only one of them needs the gear.",
+  "road.eyebrow": "Roadmap",
+  "road.h2": "Where it is.<br>Where it's going.",
+  "road.lede": "All of it is public. Each line is an issue on GitHub you can read, comment on or pick up.",
+  "road.done.h3": "Working today",
+  "road.done.i1": "Chain editor with real knobs, drag to reorder",
+  "road.done.i2": "Five backends in one rig — native, NAM, LV2, VST3, IR",
+  "road.done.i3": "Capture packs turn into knobs on their own",
+  "road.done.i4": "Presets and scenes switching without a gap in the sound",
+  "road.done.i5": "MIDI Learn and footswitch profiles",
+  "road.done.i6": "Looper, DI loop, metronome, tuner, spectrum, Tone Doctor",
+  "road.done.i7": "MCP server and the Claude skills",
+  "road.done.i8": "Several interfaces at once, insert loop for hardware",
+  "road.done.i9": "macOS build, signed and notarized",
+  "road.next.h3": "Next up",
+  "road.next.i1": "Linux and Windows builds",
+  "road.next.i2": "Setlist mode for the gig",
+  "road.next.i3": "A/B compare between two tones",
+  "road.next.i4": "Global tap tempo and expression pedal",
+  "road.next.i5": "Trim and reuse a recorded loop",
+  "road.next.i6": "Footswitch profile editor inside the app",
+  "road.next.i7": "Bass and keys rigs",
+  "road.later.h3": "Further out",
+  "road.later.i1": "Parallel routing and chain splits",
+  "road.later.i2": "Master mixer per stream",
+  "road.later.i3": "Node-graph chain editor",
+  "road.later.i4": "Backing tracks in the app",
+  "road.later.i5": "More native DSP modelled from papers",
+  "road.later.i6": "One-click plugin pack installer",
+  "road.note": "Want something moved up the list? <a href=\"https://github.com/jpfaria/OpenRig/issues\" style=\"color:var(--amber);text-decoration:underline\">Open an issue</a>.",
+  "fund.eyebrow": "Back the project",
+  "fund.h2": "Help keep it free.",
+  "fund.lede": "OpenRig has no company behind it and nothing to sell you. It stays free and open source either way — money decides how fast it moves and how many machines it runs on.",
+  "fund.c1.h3": "Hardware to test on",
+  "fund.c1.p": "Audio interfaces, footswitches and a Windows and a Linux machine to build and test on. Right now it's whatever is already in the room.",
+  "fund.c2.h3": "More gear in the catalog",
+  "fund.c2.p": "Capture packs, licences and studio time to model amps properly instead of guessing.",
+  "fund.c3.h3": "Time on the roadmap",
+  "fund.c3.p": "Every hour funded is an hour on the list above.",
+  "fund.cta1": "Sponsor on GitHub",
+  "fund.cta2": "Or contribute code",
+  "fund.note": "No tiers, no paywalled features, no \"pro\" version. Ever.",
+  "dl.eyebrow": "Get it",
+  "dl.h2": "Download now.<br>Play tonight.",
+  "dl.lede": "OpenRig is free and open source. Download it, open it and start playing — nothing to sign up for, nothing to unlock later.",
+  "dl.cta1": "Download for macOS",
+  "dl.cta2": "Read the guide",
+  "dl.meta": "v{version} · Apple silicon &amp; Intel · macOS 12 or newer · GPL-3.0",
+  "dl.changelog": "See what changed in this release",
+  "dl.betaLabel": "Feeling brave? Try the v{betaVersion} beta",
+  "dl.downloads": "{downloads} downloads so far",
+  "dl.soon": "Windows &amp; Linux — on the way"
+}
