+++ site/i18n/en/features.json
@@ -1,0 +1,38 @@
+{
+  "features.eyebrow": "Everything you need",
+  "features.h2": "More than<br>a rig.",
+  "features.lede": "Looper, tuner, metronome, spectrum analyzer and a tone diagnosis, all built in. You never leave the app to check or fix the sound.",
+  "features.f1.h3": "Looper",
+  "features.f1.p": "Record a take, listen back, record another on top. One take at a time, no undo maze.",
+  "features.f2.h3": "DI loop",
+  "features.f2.p": "Play a dry DI file through the live chain — every pedal and amp treats it exactly like your guitar. Test a tone without picking up the instrument.",
+  "features.f3.h3": "Metronome",
+  "features.f3.p": "Tap tempo, time signature, subdivision and its own output. Keeps the whole band in time.",
+  "features.f4.h3": "Tuner",
+  "features.f4.p": "A chromatic tuner per input, muting the output while you tune.",
+  "features.f5.h3": "Spectrum",
+  "features.f5.p": "Real-time spectrum per input, left and right — you can see what you're playing.",
+  "features.f6.h3": "Tone Doctor",
+  "features.f6.p": "Points out fizz, mud and boom against a genre reference, and tells you how to fix it.",
+  "features.more.presets.b": "Presets",
+  "features.more.presets.p": "Your whole board, saved. Call it back from the screen, a footswitch or Claude.",
+  "features.more.scenes.b": "Scenes",
+  "features.more.scenes.p": "Verse, chorus, solo. Same rig, one tap between them.",
+  "features.more.midi.b": "Your foot, not your mouse",
+  "features.more.midi.p": "Assign any knob or switch with MIDI Learn. Plug in an M-Vave Chocolate and it comes mapped.",
+  "features.more.multi.b": "Two interfaces at once",
+  "features.more.multi.p": "Guitar into one interface, monitors out of another. OpenRig runs both at the same time.",
+  "features.more.insert.b": "Real pedals welcome",
+  "features.more.insert.p": "Got one you won't give up? Send the signal out to it and back, mid-chain.",
+  "ai.eyebrow": "First on MCP",
+  "ai.h2": "Ask for the tone.<br>Claude builds it.",
+  "ai.lede": "OpenRig speaks MCP. Name a song or an artist and Claude looks up which amp, cab and pedals were used, picks the closest match in the catalog and saves the preset onto your rig — while you keep playing.",
+  "ai.c1.h3": "Tone Builder",
+  "ai.c1.p": "Name a song or an artist. It researches the chain that was actually used, maps it to the catalog and saves it as a new preset.",
+  "ai.c2.h3": "Grow the catalog",
+  "ai.c2.p": "Point it at your <code>.nam</code>/IR captures or a tone3000.com pack and it packages a real OpenRig plugin, manifest and all.",
+  "ai.c3.h3": "MIDI Profile Builder",
+  "ai.c3.p": "Describe your footswitch, or hand it a MIDI capture, and it writes the profile — no factory support needed.",
+  "ai.installLabel": "Install",
+  "ai.skillsText": "Six skills, one plugin — <a href=\"https://github.com/jpfaria/OpenRig-claude\" style=\"color:var(--amber);text-decoration:underline\">see them all on GitHub</a>."
+}
