+++ site/i18n/en/gear.json
@@ -1,0 +1,70 @@
+{
+  "tone.h2": "The amp responds<br>like the real one.",
+  "tone.lede": "Pick attack, sustain, the way a note compresses when you dig in. OpenRig has native amp models written in Rust and it also loads neural captures of real amps — you pick which. Either way there's no latency you can feel between your hand and the speaker.",
+  "stats.gear": "Pieces of gear",
+  "stats.native": "Native models",
+  "stats.namCapturesHeadline": "NAM captures",
+  "stats.pluginsLabel": "plugins",
+  "stats.lv2": "LV2 plugins",
+  "stats.vst3": "VST3 plugins",
+  "stats.ir": "IRs",
+  "captures.eyebrow": "Captures",
+  "captures.h2": "A capture pack<br>becomes one amp.",
+  "captures.lede": "A capture pack ships dozens of .nam files — one per gain setting, channel and mic. Everywhere else you open them one at a time. OpenRig reads the whole pack and turns it into knobs: turn Gain and the right capture loads itself.",
+  "captures.pileH": "What you downloaded",
+  "captures.pileMore": "… 31 more",
+  "captures.cap1.h3": "Turn a knob, not a list",
+  "captures.cap1.p": "Move the Gain knob and the matching capture loads underneath. It behaves like the amp, not like a file browser.",
+  "captures.cap2.h3": "Native, NAM, LV2, VST3, IR",
+  "captures.cap2.p": "Rust DSP built in, neural amp captures, LV2 and VST3 plugins and impulse responses — all five load into the same chain.",
+  "captures.cap3.h3": "Your own captures too",
+  "captures.cap3.p": "Point it at your folder of <code>.nam</code> files or IRs and it builds the same knob panel from them.",
+  "rig.eyebrow": "The rig",
+  "rig.h2": "Knobs you<br>actually turn.",
+  "rig.lede": "No menus, no parameter lists. Pedals, amps and cabs sit in a row like a pedalboard, and the knob you want is where you expect it.",
+  "rig.cap1.h3": "Grab and turn",
+  "rig.cap1.p": "Drag a knob and hear it move. Every control responds the moment you touch it.",
+  "rig.cap2.h3": "Build your board",
+  "rig.cap2.p": "Pedals in front, cabs behind, drag anything to reorder it.",
+  "rig.cap3.h3": "Save the tone",
+  "rig.cap3.p": "Found it? Save the preset. It's still there at soundcheck next month.",
+  "gearstats.amps": "Amps",
+  "gearstats.preamps": "Preamps",
+  "gearstats.cabs": "Cabs",
+  "gearstats.gain": "Drives &amp; distortion",
+  "gearstats.reverb": "Reverb",
+  "gearstats.delay": "Delay",
+  "gearstats.modulation": "Modulation",
+  "gearstats.dynamics": "Compressors",
+  "gearstats.filter": "EQ &amp; filters",
+  "gearstats.pitch": "Pitch",
+  "gearstats.body": "Acoustic bodies",
+  "gearstats.wah": "Wah",
+  "band.h2": "The whole band.<br>One machine.",
+  "band.lede": "Guitar, acoustic and a vocal mic at the same time — each with its own chain, its own gear and its own outputs. Nothing bleeds between them.",
+  "inst.eyebrow": "Not only guitar",
+  "inst.h2": "Whatever you<br>plug in.",
+  "inst.lede": "Pick the instrument and OpenRig shows only the gear that fits it.",
+  "inst.i1.b": "Electric guitar",
+  "inst.i1.p": "Amps, cabs, drives, the whole board.",
+  "inst.i2.b": "Acoustic guitar",
+  "inst.i2.p": "Body resonance, mic chain, room.",
+  "inst.i3.b": "Vocals",
+  "inst.i3.p": "Mic chain, compression, room.",
+  "inst.i4.b": "Anything else",
+  "inst.i4.p": "A generic rig for whatever the session needs.",
+  "inst.note": "Bass and keys rigs are next up.",
+  "gig.eyebrow": "Built for the gig",
+  "gig.h2": "It holds up<br>on stage.",
+  "gig.lede": "Rehearsal room, living room or a stage with a click in your in-ears — it behaves the same every night.",
+  "gig.c1.h3": "Latency you can't feel",
+  "gig.c1.p": "The signal goes out and comes back fast enough that your hands never notice the computer.",
+  "gig.c2.h3": "Switch without a click",
+  "gig.c2.p": "Change amp, add a pedal or load a preset while you're playing. No gap, no pop, no dead bar.",
+  "gig.c3.h3": "No amp head, no pedalboard",
+  "gig.c3.p": "You already have an audio interface to plug the guitar into. Everything after it — amp, cab, pedals — runs on the laptop.",
+  "app.eyebrow": "Inside",
+  "app.h2": "Everything<br>in one view.",
+  "app.lede": "Instruments on the left, gear in the middle, outputs on the right. That's the whole app.",
+  "app.caption": "Every knob, every parameter — one screen."
+}
