+++ site/i18n/en/shell.json
@@ -1,0 +1,40 @@
+{
+  "nav.tone": "Tone",
+  "nav.captures": "Captures",
+  "nav.rig": "The rig",
+  "nav.band": "Band",
+  "nav.features": "Features",
+  "nav.ai": "Claude",
+  "nav.app": "Inside",
+  "nav.roadmap": "Roadmap",
+  "nav.support": "Support",
+  "nav.docs": "Docs",
+  "nav.download": "Download",
+  "hero.eyebrow": "Free · Open source · Mac",
+  "hero.h1": "Your amps.<br>Your pedals.<br><em>Your laptop.</em>",
+  "hero.lede": "{gear} amps, cabs and pedals running on the laptop you already own. Plug your guitar into your interface and play — free, with no latency you can feel.",
+  "hero.cta1": "Download for Mac",
+  "hero.cta2": "See the rig",
+  "hero.note": "Free forever. No account, no cloud, no subscription.",
+  "hero.scroll": "Scroll",
+  "hero.supportBig": "Back the project",
+  "hero.supportSub": "GitHub Sponsors",
+  "hero.supportBtn": "♥ Back the project",
+  "footer.learn.h4": "Learn",
+  "footer.learn.l1": "Getting started",
+  "footer.learn.l2": "Building a rig",
+  "footer.learn.l3": "Presets",
+  "footer.learn.l4": "Audio setup",
+  "footer.gear.h4": "Gear",
+  "footer.gear.l1": "Amps",
+  "footer.gear.l2": "Cabinets",
+  "footer.gear.l3": "Pedals",
+  "footer.gear.l4": "Full catalog",
+  "footer.project.h4": "Project",
+  "footer.project.l1": "GitHub",
+  "footer.project.l2": "Releases",
+  "footer.project.l3": "Report a bug",
+  "footer.project.l4": "Contribute",
+  "footer.project.l5": "Sponsor",
+  "footer.copyright": "© 2026 OpenRig · GPL-3.0"
+}
