+++ site/i18n/es-ES/end.json
@@ -1,0 +1,59 @@
+{
+  "story.eyebrow": "De dónde viene",
+  "story.h2": "Esto empezó porque tres<br>pedaleras cuestan una fortuna.",
+  "story.p1": "Mi hijo empezó clases de guitarra. El profesor tenía un sonido, nosotros teníamos otro en casa, y lo que aprendía en la clase no se repetía después. Resolverlo comprando la misma pedalera para todos costaba más que el instrumento.",
+  "story.p2": "Llevo 26 años como desarrollador. En ese momento estaba investigando, con método, hasta dónde se puede llegar programando con IA — y un estudio así solo vale si el problema es duro. El audio en vivo es duro: el código corre unos cientos de veces por segundo y, cuando se equivoca, el sonido falla al instante, delante de todo el mundo. No hay forma de disimularlo.",
+  "story.p3": "Salió mejor de lo que esperaba. OpenRig hoy es una plataforma de audio en tiempo real escrita en Rust: cada entrada es un runtime aislado, así que dos guitarras en la misma interfaz son de verdad dos rigs independientes. Ampli, bafle y pedal mezclando DSP nativo, capturas neuronales e IRs en la misma cadena. Presets en texto plano que puedes abrir y editar.",
+  "story.p4": "Y la idea original creció con él: en vez de un timbre para mi hijo y para mí, un ordenador en el escenario procesando a toda la banda, cada músico ajustando su propio sonido. Solo uno necesita el equipo.",
+  "road.eyebrow": "Roadmap",
+  "road.h2": "Dónde está.<br>Hacia dónde va.",
+  "road.lede": "Todo es público. Cada línea es una issue en GitHub que puedes leer, comentar o tomar.",
+  "road.done.h3": "Funcionando hoy",
+  "road.done.i1": "Editor de cadena con knobs de verdad, arrastra para reordenar",
+  "road.done.i2": "Cinco backends en el mismo rig — nativo, NAM, LV2, VST3, IR",
+  "road.done.i3": "Los packs de capturas se vuelven knobs solos",
+  "road.done.i4": "Presets y scenes cambiando sin un hueco en el sonido",
+  "road.done.i5": "MIDI Learn y perfiles de footswitch",
+  "road.done.i6": "Looper, DI loop, metrónomo, afinador, espectro, Tone Doctor",
+  "road.done.i7": "Servidor MCP y las skills de Claude",
+  "road.done.i8": "Varias interfaces a la vez, insert para hardware",
+  "road.done.i9": "Build de macOS, firmado y notarizado",
+  "road.next.h3": "Lo siguiente",
+  "road.next.i1": "Builds de Linux y Windows",
+  "road.next.i2": "Modo setlist para el show",
+  "road.next.i3": "Comparar dos timbres en A/B",
+  "road.next.i4": "Tap tempo global y pedal de expresión",
+  "road.next.i5": "Recortar y reutilizar un loop grabado",
+  "road.next.i6": "Editor de perfil de footswitch dentro de la app",
+  "road.next.i7": "Rigs de bajo y teclado",
+  "road.later.h3": "Más adelante",
+  "road.later.i1": "Ruteo paralelo y divisiones de cadena",
+  "road.later.i2": "Mixer master por stream",
+  "road.later.i3": "Editor de cadena en grafo de nodos",
+  "road.later.i4": "Pistas de acompañamiento dentro de la app",
+  "road.later.i5": "Más DSP nativo modelado desde papers",
+  "road.later.i6": "Instalador de packs de plugins en un clic",
+  "road.note": "¿Quieres algo más arriba en la lista? <a href=\"https://github.com/jpfaria/OpenRig/issues\" style=\"color:var(--amber);text-decoration:underline\">Abre una issue</a>.",
+  "fund.eyebrow": "Apoya el proyecto",
+  "fund.h2": "Ayuda a mantenerlo<br>gratis.",
+  "fund.lede": "OpenRig no tiene una empresa detrás ni nada que venderte. Sigue siendo gratis y de código abierto de todos modos — el dinero decide la velocidad y en cuántas máquinas corre.",
+  "fund.c1.h3": "Hardware para probar",
+  "fund.c1.p": "Interfaces de audio, footswitches y una máquina Windows y una Linux para compilar y probar. Hoy es solo lo que ya hay en casa.",
+  "fund.c2.h3": "Más equipo en el catálogo",
+  "fund.c2.p": "Packs de capturas, licencias y tiempo de estudio para modelar amplis en serio en vez de adivinar.",
+  "fund.c3.h3": "Tiempo en el roadmap",
+  "fund.c3.p": "Cada hora financiada es una hora en la lista de arriba.",
+  "fund.cta1": "Apoyar en GitHub",
+  "fund.cta2": "O contribuir con código",
+  "fund.note": "Sin planes, sin funciones de pago, sin versión \"pro\". Nunca.",
+  "dl.eyebrow": "Consíguelo",
+  "dl.h2": "Descarga ahora.<br>Toca hoy.",
+  "dl.lede": "OpenRig es gratis y de código abierto. Descárgalo, ábrelo y empieza a tocar — nada que registrar, nada que desbloquear después.",
+  "dl.cta1": "Descargar para macOS",
+  "dl.cta2": "Leer la guía",
+  "dl.meta": "v{version} · Apple silicon &amp; Intel · macOS 12 o más nuevo · GPL-3.0",
+  "dl.changelog": "Ve qué cambió en esta versión",
+  "dl.betaLabel": "¿Te arriesgas? Prueba la beta v{betaVersion}",
+  "dl.downloads": "{downloads} descargas hasta ahora",
+  "dl.soon": "Windows &amp; Linux — en camino"
+}
