+++ site/i18n/es-ES/features.json
@@ -1,0 +1,38 @@
+{
+  "features.eyebrow": "Todo lo que necesitas",
+  "features.h2": "Más que<br>un rig.",
+  "features.lede": "Looper, afinador, metrónomo, analizador de espectro y un diagnóstico de timbre, todo integrado. No sales de la app ni para comprobar ni para corregir el sonido.",
+  "features.f1.h3": "Looper",
+  "features.f1.p": "Graba una toma, escúchala, graba otra encima. Una toma a la vez, sin laberinto de deshacer.",
+  "features.f2.h3": "DI loop",
+  "features.f2.p": "Reproduce un DI seco por la cadena en vivo — cada pedal y cada ampli lo trata igual que a tu guitarra. Prueba un timbre sin coger el instrumento.",
+  "features.f3.h3": "Metrónomo",
+  "features.f3.p": "Tap tempo, compás, subdivisión y salida propia. Mantiene a toda la banda en tiempo.",
+  "features.f4.h3": "Afinador",
+  "features.f4.p": "Un afinador cromático por entrada, silenciando la salida mientras afinas.",
+  "features.f5.h3": "Espectro",
+  "features.f5.p": "Espectro en tiempo real por entrada, izquierda y derecha — se ve lo que estás tocando.",
+  "features.f6.h3": "Tone Doctor",
+  "features.f6.p": "Señala fizz, mud y boom comparando con una referencia del género, y dice cómo corregirlo.",
+  "features.more.presets.b": "Presets",
+  "features.more.presets.p": "Tu pedalboard entero, guardado. Llámalo de vuelta desde la pantalla, un footswitch o Claude.",
+  "features.more.scenes.b": "Scenes",
+  "features.more.scenes.p": "Verso, estribillo, solo. El mismo rig, un toque entre ellos.",
+  "features.more.midi.b": "Con el pie, no con el mouse",
+  "features.more.midi.p": "Asigna cualquier knob o interruptor con MIDI Learn. Conecta un M-Vave Chocolate y ya viene mapeado.",
+  "features.more.multi.b": "Dos interfaces a la vez",
+  "features.more.multi.p": "Guitarra entrando por una interfaz, monitores saliendo por otra. OpenRig corre las dos a la vez.",
+  "features.more.insert.b": "Pedales reales bienvenidos",
+  "features.more.insert.p": "¿Tienes uno del que no te separas? Manda la señal hacia él y de vuelta, en medio de la cadena.",
+  "ai.eyebrow": "El primero con MCP",
+  "ai.h2": "Pide el timbre.<br>Claude lo arma.",
+  "ai.lede": "OpenRig habla MCP. Dices el nombre de una canción o de un artista y Claude investiga qué ampli, pantalla y pedales se usaron, elige lo más parecido del catálogo y guarda el preset en tu rig — mientras tú sigues tocando.",
+  "ai.c1.h3": "Tone Builder",
+  "ai.c1.p": "Nombra una canción o un artista. Investiga la cadena que se usó de verdad, la mapea al catálogo y la guarda como preset nuevo.",
+  "ai.c2.h3": "Amplía el catálogo",
+  "ai.c2.p": "Apúntalo a tus capturas <code>.nam</code>/IR o a un pack de tone3000.com y empaqueta un plugin de OpenRig de verdad, con manifest y todo.",
+  "ai.c3.h3": "MIDI Profile Builder",
+  "ai.c3.p": "Describe tu footswitch, o dale una captura MIDI, y escribe el perfil — no hace falta que el fabricante lo soporte.",
+  "ai.installLabel": "Instalar",
+  "ai.skillsText": "Seis skills, un solo plugin — <a href=\"https://github.com/jpfaria/OpenRig-claude\" style=\"color:var(--amber);text-decoration:underline\">míralas todas en GitHub</a>."
+}
