+++ site/i18n/es-ES/gear.json
@@ -1,0 +1,70 @@
+{
+  "tone.h2": "El ampli responde<br>como el de verdad.",
+  "tone.lede": "Ataque de púa, sustain, cómo comprime la nota cuando aprietas. OpenRig tiene amplis nativos escritos en Rust y también carga capturas neuronales de amplis reales — tú eliges. En ambos casos, sin latencia que puedas sentir entre tu mano y el altavoz.",
+  "stats.gear": "Equipos",
+  "stats.native": "Modelos nativos",
+  "stats.namCapturesHeadline": "Capturas NAM",
+  "stats.pluginsLabel": "plugins",
+  "stats.lv2": "Plugins LV2",
+  "stats.vst3": "Plugins VST3",
+  "stats.ir": "IRs",
+  "captures.eyebrow": "Capturas",
+  "captures.h2": "Un pack de capturas<br>se vuelve un ampli.",
+  "captures.lede": "Un pack de capturas trae decenas de .nam — uno por cada ganancia, canal y micrófono. En los otros apps los abres uno por uno. OpenRig lee el pack entero y lo convierte en knobs: giras el Gain y la captura correcta entra sola.",
+  "captures.pileH": "Lo que descargaste",
+  "captures.pileMore": "… 31 más",
+  "captures.cap1.h3": "Gira un knob, no una lista",
+  "captures.cap1.p": "Mueves el knob de Gain y la captura correspondiente entra por debajo. Responde como el ampli, no como un explorador de archivos.",
+  "captures.cap2.h3": "Nativo, NAM, LV2, VST3, IR",
+  "captures.cap2.p": "DSP en Rust integrado, capturas neuronales de amplis, plugins LV2 y VST3 e impulse responses — los cinco entran en la misma cadena.",
+  "captures.cap3.h3": "Tus propias capturas también",
+  "captures.cap3.p": "Apúntalo a tu carpeta de <code>.nam</code> o de IRs y arma el mismo panel de knobs a partir de ellas.",
+  "rig.eyebrow": "El rig",
+  "rig.h2": "Knobs que<br>realmente giras.",
+  "rig.lede": "Sin menús, sin listas de parámetros. Pedal, ampli y bafle se alinean como un pedalboard, y el knob que buscas está donde lo esperas.",
+  "rig.cap1.h3": "Agarra y gira",
+  "rig.cap1.p": "Arrastra el knob y escúchalo cambiar. Cada control responde en el instante en que lo tocas.",
+  "rig.cap2.h3": "Arma tu pedalboard",
+  "rig.cap2.p": "Pedales delante, bafles detrás, y arrastra para reordenar cualquiera.",
+  "rig.cap3.h3": "Guarda el timbre",
+  "rig.cap3.p": "¿Lo encontraste? Guarda el preset. Seguirá ahí en la prueba de sonido del mes que viene.",
+  "gearstats.amps": "Amplis",
+  "gearstats.preamps": "Preamplis",
+  "gearstats.cabs": "Bafles",
+  "gearstats.gain": "Drives &amp; distorsiones",
+  "gearstats.reverb": "Reverb",
+  "gearstats.delay": "Delay",
+  "gearstats.modulation": "Modulación",
+  "gearstats.dynamics": "Compresores",
+  "gearstats.filter": "EQ &amp; filtros",
+  "gearstats.pitch": "Pitch",
+  "gearstats.body": "Cuerpos acústicos",
+  "gearstats.wah": "Wah",
+  "band.h2": "Toda la banda.<br>Una sola máquina.",
+  "band.lede": "Guitarra, acústica y un micrófono de voz a la vez — cada uno con su cadena, su equipo y sus salidas. Nada se filtra de uno a otro.",
+  "inst.eyebrow": "No solo guitarra",
+  "inst.h2": "Lo que sea<br>que conectes.",
+  "inst.lede": "Elige el instrumento y OpenRig muestra solo el equipo que le sirve.",
+  "inst.i1.b": "Guitarra eléctrica",
+  "inst.i1.p": "Amplis, bafles, drives, todo el pedalboard.",
+  "inst.i2.b": "Guitarra acústica",
+  "inst.i2.p": "Resonancia del cuerpo, cadena de micrófono, ambiente.",
+  "inst.i3.b": "Voz",
+  "inst.i3.p": "Cadena de micrófono, compresión, ambiente.",
+  "inst.i4.b": "Cualquier otra cosa",
+  "inst.i4.p": "Un rig genérico para lo que pida la sesión.",
+  "inst.note": "Bajo y teclado son los siguientes.",
+  "gig.eyebrow": "Hecho para el show",
+  "gig.h2": "Aguanta<br>el show.",
+  "gig.lede": "Sala de ensayo, salón de casa o escenario con clic en el in-ear — se comporta igual todas las noches.",
+  "gig.c1.h3": "Latencia que no sientes",
+  "gig.c1.p": "La señal sale y vuelve lo bastante rápido para que tus manos nunca noten el ordenador.",
+  "gig.c2.h3": "Cambios sin chasquido",
+  "gig.c2.p": "Cambia de ampli, pon un pedal o carga un preset mientras tocas. Sin cortes, sin chasquidos, sin compases perdidos.",
+  "gig.c3.h3": "Sin cabezal, sin pedalera",
+  "gig.c3.p": "Ya tienes la interfaz de audio para enchufar la guitarra. De ahí en adelante — ampli, pantalla, pedales — todo corre en el portátil.",
+  "app.eyebrow": "Por dentro",
+  "app.h2": "Todo en<br>una sola vista.",
+  "app.lede": "Instrumentos a la izquierda, equipo en el medio, salidas a la derecha. Esa es toda la app.",
+  "app.caption": "Cada knob, cada parámetro — una sola pantalla."
+}
