+++ site/i18n/es-ES/shell.json
@@ -1,0 +1,40 @@
+{
+  "nav.tone": "Timbre",
+  "nav.captures": "Capturas",
+  "nav.rig": "El rig",
+  "nav.band": "Banda",
+  "nav.features": "Funciones",
+  "nav.ai": "Claude",
+  "nav.app": "Por dentro",
+  "nav.roadmap": "Roadmap",
+  "nav.support": "Apoyar",
+  "nav.docs": "Docs",
+  "nav.download": "Descargar",
+  "hero.eyebrow": "Gratis · Código abierto · Mac",
+  "hero.h1": "Tus amplis.<br>Tus pedales.<br><em>Tu portátil.</em>",
+  "hero.lede": "{gear} amplis, bafles y pedales corriendo en el portátil que ya tienes. Conecta la guitarra a tu interfaz y toca — gratis, sin latencia que puedas sentir.",
+  "hero.cta1": "Descargar para Mac",
+  "hero.cta2": "Ver el rig",
+  "hero.note": "Gratis para siempre. Sin cuenta, sin nube, sin suscripción.",
+  "hero.scroll": "Baja",
+  "hero.supportBig": "Apoya el proyecto",
+  "hero.supportSub": "GitHub Sponsors",
+  "hero.supportBtn": "♥ Apoya el proyecto",
+  "footer.learn.h4": "Aprender",
+  "footer.learn.l1": "Primeros pasos",
+  "footer.learn.l2": "Armar un rig",
+  "footer.learn.l3": "Presets",
+  "footer.learn.l4": "Configuración de audio",
+  "footer.gear.h4": "Equipos",
+  "footer.gear.l1": "Amplis",
+  "footer.gear.l2": "Bafles",
+  "footer.gear.l3": "Pedales",
+  "footer.gear.l4": "Catálogo completo",
+  "footer.project.h4": "Proyecto",
+  "footer.project.l1": "GitHub",
+  "footer.project.l2": "Releases",
+  "footer.project.l3": "Reportar un error",
+  "footer.project.l4": "Contribuir",
+  "footer.project.l5": "Apoyar",
+  "footer.copyright": "© 2026 OpenRig · GPL-3.0"
+}
