+++ site/i18n/pt-BR/end.json
@@ -1,0 +1,59 @@
+{
+  "story.eyebrow": "De onde veio",
+  "story.h2": "Isso começou porque três<br>pedaleiras custam caro.",
+  "story.p1": "Meu filho começou a ter aula de guitarra. O professor tinha um som, a gente tinha outro em casa, e o que ele aprendia na aula não se repetia depois. Resolver comprando a mesma pedaleira pra todo mundo custava mais que o instrumento.",
+  "story.p2": "Sou desenvolvedor há 26 anos. Na época eu estava investigando, com método, até onde dá pra ir programando com IA — e um estudo desses só vale se o problema for duro. Áudio ao vivo é duro: o código roda algumas centenas de vezes por segundo e, quando erra, o som falha na hora, na frente de todo mundo. Não tem como maquiar.",
+  "story.p3": "Deu certo além do que eu esperava. O OpenRig hoje é uma plataforma de áudio em tempo real escrita em Rust: cada entrada é um runtime isolado, então duas guitarras na mesma interface são dois rigs independentes de verdade. Amp, caixa e pedal misturando DSP nativo, captura neural e IR na mesma cadeia. Preset em texto aberto, que você abre e edita.",
+  "story.p4": "E a ideia do começo cresceu junto: em vez de um timbre pra mim e pro meu filho, um computador no palco processando a banda inteira, cada músico mexendo no próprio som. Só um precisa do equipamento.",
+  "road.eyebrow": "Roadmap",
+  "road.h2": "Onde ele está.<br>Pra onde ele vai.",
+  "road.lede": "Tudo é público. Cada linha é uma issue no GitHub que você pode ler, comentar ou pegar pra fazer.",
+  "road.done.h3": "Funcionando hoje",
+  "road.done.i1": "Editor de cadeia com knob de verdade, arrasta pra reordenar",
+  "road.done.i2": "Cinco backends no mesmo rig — nativo, NAM, LV2, VST3, IR",
+  "road.done.i3": "Pack de captura vira knob sozinho",
+  "road.done.i4": "Presets e scenes trocando sem buraco no som",
+  "road.done.i5": "MIDI Learn e profile de footswitch",
+  "road.done.i6": "Looper, DI loop, metrônomo, afinador, espectro, Tone Doctor",
+  "road.done.i7": "Servidor MCP e as skills do Claude",
+  "road.done.i8": "Várias interfaces ao mesmo tempo, insert pra hardware",
+  "road.done.i9": "Build de macOS, assinado e notarizado",
+  "road.next.h3": "Vem a seguir",
+  "road.next.i1": "Builds de Linux e Windows",
+  "road.next.i2": "Modo setlist pro show",
+  "road.next.i3": "Comparar dois timbres em A/B",
+  "road.next.i4": "Tap tempo global e pedal de expressão",
+  "road.next.i5": "Cortar e reaproveitar um loop gravado",
+  "road.next.i6": "Editor de profile de footswitch dentro do app",
+  "road.next.i7": "Rigs de baixo e teclado",
+  "road.later.h3": "Mais pra frente",
+  "road.later.i1": "Roteamento paralelo e divisão de cadeia",
+  "road.later.i2": "Mixer master por stream",
+  "road.later.i3": "Editor de cadeia em grafo de nós",
+  "road.later.i4": "Backing tracks dentro do app",
+  "road.later.i5": "Mais DSP nativo modelado a partir de papers",
+  "road.later.i6": "Instalador de pack de plugin em um clique",
+  "road.note": "Quer alguma coisa mais pra cima na lista? <a href=\"https://github.com/jpfaria/OpenRig/issues\" style=\"color:var(--amber);text-decoration:underline\">Abre uma issue</a>.",
+  "fund.eyebrow": "Apoie o projeto",
+  "fund.h2": "Ajuda a manter<br>de graça.",
+  "fund.lede": "O OpenRig não tem empresa atrás e não tem nada pra te vender. Continua grátis e de código aberto de qualquer jeito — o dinheiro decide a velocidade e em quantas máquinas ele roda.",
+  "fund.c1.h3": "Hardware pra testar",
+  "fund.c1.p": "Interface de áudio, footswitch e uma máquina Windows e uma Linux pra buildar e testar. Hoje é só o que já tem em casa.",
+  "fund.c2.h3": "Mais equipamento no catálogo",
+  "fund.c2.p": "Pack de captura, licença e tempo de estúdio pra modelar amp direito em vez de chutar.",
+  "fund.c3.h3": "Tempo no roadmap",
+  "fund.c3.p": "Cada hora bancada é uma hora na lista aí de cima.",
+  "fund.cta1": "Apoiar no GitHub",
+  "fund.cta2": "Ou contribuir com código",
+  "fund.note": "Sem plano, sem feature trancada, sem versão \"pro\". Nunca.",
+  "dl.eyebrow": "Baixe",
+  "dl.h2": "Baixa agora.<br>Toca hoje.",
+  "dl.lede": "O OpenRig é grátis e de código aberto. Baixa, abre e começa a tocar — nada pra assinar, nada pra desbloquear depois.",
+  "dl.cta1": "Baixar para macOS",
+  "dl.cta2": "Ler o guia",
+  "dl.meta": "v{version} · Apple silicon &amp; Intel · macOS 12 ou mais novo · GPL-3.0",
+  "dl.changelog": "Veja o que mudou nesta versão",
+  "dl.betaLabel": "Quer arriscar? Testa a beta v{betaVersion}",
+  "dl.downloads": "{downloads} downloads até agora",
+  "dl.soon": "Windows &amp; Linux — a caminho"
+}
