+++ site/i18n/pt-BR/features.json
@@ -1,0 +1,38 @@
+{
+  "features.eyebrow": "Tudo que você precisa",
+  "features.h2": "Mais do que<br>um rig.",
+  "features.lede": "Looper, afinador, metrônomo, analisador de espectro e um diagnóstico de timbre, tudo embutido. Você não sai do app pra conferir nem pra corrigir o som.",
+  "features.f1.h3": "Looper",
+  "features.f1.p": "Grava um take, escuta, grava outro por cima. Um take por vez, sem labirinto de desfazer.",
+  "features.f2.h3": "DI loop",
+  "features.f2.p": "Toca um DI seco pela cadeia ao vivo — cada pedal e cada amp trata igualzinho à tua guitarra. Testa timbre sem pegar o instrumento.",
+  "features.f3.h3": "Metrônomo",
+  "features.f3.p": "Tap tempo, fórmula de compasso, subdivisão e saída própria. Segura a banda inteira no tempo.",
+  "features.f4.h3": "Afinador",
+  "features.f4.p": "Um afinador cromático por entrada, mutando a saída enquanto você afina.",
+  "features.f5.h3": "Espectro",
+  "features.f5.p": "Espectro em tempo real por entrada, esquerda e direita — dá pra ver o que você está tocando.",
+  "features.f6.h3": "Tone Doctor",
+  "features.f6.p": "Aponta fizz, mud e boom comparando com uma referência do gênero, e diz como corrigir.",
+  "features.more.presets.b": "Presets",
+  "features.more.presets.p": "Teu board inteiro, salvo. Chama de volta pela tela, pelo footswitch ou pelo Claude.",
+  "features.more.scenes.b": "Scenes",
+  "features.more.scenes.p": "Verso, refrão, solo. Mesmo rig, um toque entre eles.",
+  "features.more.midi.b": "No pé, não no mouse",
+  "features.more.midi.p": "Atribui qualquer knob ou chave com MIDI Learn. Pluga um M-Vave Chocolate e já vem mapeado.",
+  "features.more.multi.b": "Duas interfaces ao mesmo tempo",
+  "features.more.multi.p": "Guitarra entrando numa interface, monitor saindo na outra. O OpenRig roda as duas ao mesmo tempo.",
+  "features.more.insert.b": "Pedal de verdade é bem-vindo",
+  "features.more.insert.p": "Tem um que você não abre mão? Manda o sinal pra ele e traz de volta, no meio da cadeia.",
+  "ai.eyebrow": "Primeiro no MCP",
+  "ai.h2": "Peça o timbre.<br>O Claude monta.",
+  "ai.lede": "O OpenRig fala MCP. Você diz o nome de uma música ou de um artista e o Claude pesquisa qual amp, caixa e pedais foram usados, escolhe o mais próximo no catálogo e salva o preset no teu rig — enquanto você continua tocando.",
+  "ai.c1.h3": "Tone Builder",
+  "ai.c1.p": "Diz o nome de uma música ou de um artista. Ele pesquisa a cadeia que foi usada de verdade, mapeia pro catálogo e salva como preset novo.",
+  "ai.c2.h3": "Aumenta o catálogo",
+  "ai.c2.p": "Aponta pras tuas capturas <code>.nam</code>/IR ou pra um pack do tone3000.com e ele empacota um plugin OpenRig de verdade, com manifest e tudo.",
+  "ai.c3.h3": "MIDI Profile Builder",
+  "ai.c3.p": "Descreve o teu footswitch, ou entrega uma captura MIDI, e ele escreve o profile — não precisa o fabricante ter suporte.",
+  "ai.installLabel": "Instalar",
+  "ai.skillsText": "Seis skills, um plugin só — <a href=\"https://github.com/jpfaria/OpenRig-claude\" style=\"color:var(--amber);text-decoration:underline\">veja todas no GitHub</a>."
+}
