+++ site/i18n/pt-BR/gear.json
@@ -1,0 +1,70 @@
+{
+  "tone.h2": "O amp responde<br>como o de verdade.",
+  "tone.lede": "Ataque de palheta, sustain, o jeito que a nota comprime quando você força. O OpenRig tem amp nativo escrito em Rust e também carrega captura neural de amp real — você escolhe. Nos dois casos, sem latência que você sinta entre a mão e o alto-falante.",
+  "stats.gear": "Equipamentos",
+  "stats.native": "Modelos nativos",
+  "stats.namCapturesHeadline": "Capturas NAM",
+  "stats.pluginsLabel": "plugins",
+  "stats.lv2": "Plugins LV2",
+  "stats.vst3": "Plugins VST3",
+  "stats.ir": "IRs",
+  "captures.eyebrow": "Capturas",
+  "captures.h2": "Um pacote de capturas<br>vira um amp só.",
+  "captures.lede": "Pack de captura vem com dezenas de .nam — um pra cada ganho, canal e microfone. Nos outros apps você abre um por um. O OpenRig lê o pack inteiro e transforma em knob: gira o Gain e a captura certa entra sozinha.",
+  "captures.pileH": "O que você baixou",
+  "captures.pileMore": "… mais 31",
+  "captures.cap1.h3": "Gira um knob, não uma lista",
+  "captures.cap1.p": "Move o knob de Gain e a captura correspondente entra por baixo. Se comporta como o amp, não como navegador de arquivo.",
+  "captures.cap2.h3": "Nativo, NAM, LV2, VST3, IR",
+  "captures.cap2.p": "DSP em Rust embutido, captura neural de amp, plugins LV2 e VST3 e impulse responses — os cinco entram na mesma cadeia.",
+  "captures.cap3.h3": "As tuas capturas também",
+  "captures.cap3.p": "Aponta pra tua pasta de <code>.nam</code> ou de IRs e ele monta o mesmo painel de knobs a partir delas.",
+  "rig.eyebrow": "O rig",
+  "rig.h2": "Knobs que você<br>realmente gira.",
+  "rig.lede": "Sem menu, sem lista de parâmetro. Pedal, amp e caixa ficam enfileirados como um pedalboard, e o knob que você quer está onde você espera.",
+  "rig.cap1.h3": "Pega e gira",
+  "rig.cap1.p": "Arrasta o knob e escuta mudar. Todo controle responde no instante em que você encosta.",
+  "rig.cap2.h3": "Monta o teu board",
+  "rig.cap2.p": "Pedal na frente, caixa atrás, e arrasta pra reordenar qualquer um.",
+  "rig.cap3.h3": "Salva o timbre",
+  "rig.cap3.p": "Achou? Salva o preset. Ele vai estar lá na passagem de som do mês que vem.",
+  "gearstats.amps": "Amps",
+  "gearstats.preamps": "Pré-amps",
+  "gearstats.cabs": "Caixas",
+  "gearstats.gain": "Drives &amp; distorções",
+  "gearstats.reverb": "Reverb",
+  "gearstats.delay": "Delay",
+  "gearstats.modulation": "Modulação",
+  "gearstats.dynamics": "Compressores",
+  "gearstats.filter": "EQ &amp; filtros",
+  "gearstats.pitch": "Pitch",
+  "gearstats.body": "Corpos de violão",
+  "gearstats.wah": "Wah",
+  "band.h2": "A banda inteira.<br>Uma máquina só.",
+  "band.lede": "Guitarra, violão e um microfone de voz ao mesmo tempo — cada um com a sua cadeia, o seu equipamento e as suas saídas. Nada vaza de um pro outro.",
+  "inst.eyebrow": "Não só guitarra",
+  "inst.h2": "O que você<br>plugar.",
+  "inst.lede": "Escolhe o instrumento e o OpenRig mostra só o equipamento que serve pra ele.",
+  "inst.i1.b": "Guitarra",
+  "inst.i1.p": "Amps, caixas, drives, o board inteiro.",
+  "inst.i2.b": "Violão",
+  "inst.i2.p": "Ressonância do corpo, cadeia de microfone, ambiência.",
+  "inst.i3.b": "Voz",
+  "inst.i3.p": "Cadeia de microfone, compressão, ambiência.",
+  "inst.i4.b": "Qualquer outra coisa",
+  "inst.i4.p": "Um rig genérico pro que a sessão pedir.",
+  "inst.note": "Baixo e teclado são os próximos.",
+  "gig.eyebrow": "Feito pro show",
+  "gig.h2": "Aguenta<br>o show.",
+  "gig.lede": "Sala de ensaio, sala de casa ou palco com clique no in-ear — se comporta igual toda noite.",
+  "gig.c1.h3": "Latência que você não sente",
+  "gig.c1.p": "O sinal sai e volta rápido o suficiente pra tua mão nunca perceber que tem um computador no meio.",
+  "gig.c2.h3": "Troca sem estalo",
+  "gig.c2.p": "Troca de amp, coloca um pedal ou carrega um preset tocando. Sem buraco, sem estalo, sem compasso perdido.",
+  "gig.c3.h3": "Sem cabeçote, sem pedaleira",
+  "gig.c3.p": "Você já tem a interface de áudio pra plugar a guitarra. Do plug em diante — amp, caixa, pedais — roda tudo no notebook.",
+  "app.eyebrow": "Por dentro",
+  "app.h2": "Tudo em<br>uma tela só.",
+  "app.lede": "Instrumentos na esquerda, equipamento no meio, saídas na direita. É o app inteiro.",
+  "app.caption": "Cada knob, cada parâmetro — uma tela só."
+}
