+++ site/i18n/pt-BR/shell.json
@@ -1,0 +1,40 @@
+{
+  "nav.tone": "Timbre",
+  "nav.captures": "Capturas",
+  "nav.rig": "O rig",
+  "nav.band": "Banda",
+  "nav.features": "Recursos",
+  "nav.ai": "Claude",
+  "nav.app": "Por dentro",
+  "nav.roadmap": "Roadmap",
+  "nav.support": "Apoiar",
+  "nav.docs": "Docs",
+  "nav.download": "Baixar",
+  "hero.eyebrow": "Grátis · Código aberto · Mac",
+  "hero.h1": "Seus amps.<br>Seus pedais.<br><em>Seu laptop.</em>",
+  "hero.lede": "{gear} amps, caixas e pedais rodando no notebook que você já tem. Pluga a guitarra na interface e toca — de graça, sem latência que você sinta.",
+  "hero.cta1": "Baixar para Mac",
+  "hero.cta2": "Ver o rig",
+  "hero.note": "Grátis pra sempre. Sem conta, sem nuvem, sem assinatura.",
+  "hero.scroll": "Desce",
+  "hero.supportBig": "Apoie o projeto",
+  "hero.supportSub": "GitHub Sponsors",
+  "hero.supportBtn": "♥ Apoie o projeto",
+  "footer.learn.h4": "Aprender",
+  "footer.learn.l1": "Primeiros passos",
+  "footer.learn.l2": "Montando um rig",
+  "footer.learn.l3": "Presets",
+  "footer.learn.l4": "Configuração de áudio",
+  "footer.gear.h4": "Equipamentos",
+  "footer.gear.l1": "Amps",
+  "footer.gear.l2": "Caixas",
+  "footer.gear.l3": "Pedais",
+  "footer.gear.l4": "Catálogo completo",
+  "footer.project.h4": "Projeto",
+  "footer.project.l1": "GitHub",
+  "footer.project.l2": "Releases",
+  "footer.project.l3": "Reportar um bug",
+  "footer.project.l4": "Contribuir",
+  "footer.project.l5": "Apoiar",
+  "footer.copyright": "© 2026 OpenRig · GPL-3.0"
+}
