+++ site/index.html
@@ -8,7 +8,25 @@
 <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Archivo:ital,wght@0,300;0,400;0,500;0,600;0,700;1,400&display=swap" rel="stylesheet">
-<link rel="stylesheet" href="styles.css">
+<link rel="stylesheet" href="css/tokens.css">
+<link rel="stylesheet" href="css/base.css">
+<link rel="stylesheet" href="css/type.css">
+<link rel="stylesheet" href="css/buttons.css">
+<link rel="stylesheet" href="css/nav.css">
+<link rel="stylesheet" href="css/hero.css">
+<link rel="stylesheet" href="css/layout.css">
+<link rel="stylesheet" href="css/band.css">
+<link rel="stylesheet" href="css/gear.css">
+<link rel="stylesheet" href="css/stats.css">
+<link rel="stylesheet" href="css/captures.css">
+<link rel="stylesheet" href="css/cards.css">
+<link rel="stylesheet" href="css/features.css">
+<link rel="stylesheet" href="css/roadmap.css">
+<link rel="stylesheet" href="css/story.css">
+<link rel="stylesheet" href="css/download.css">
+<link rel="stylesheet" href="css/footer.css">
+<!-- last: the cross-section phone rules have to win over each section's own file -->
+<link rel="stylesheet" href="css/responsive.css">
 </head>
 <body>
 
@@ -44,25 +62,25 @@
     <img src="assets/feat-project.png" alt="OpenRig main window — amps and pedals laid out like a pedalboard">
     <img class="hero-product-2" src="assets/feat-chain.png" alt="OpenRig chain view — the knobs you turn while playing">
     <a class="hero-support" href="#support">
-      <span data-i18n="hero.supportBig">Back OpenRig</span>
-      <i data-i18n="hero.supportSub">Free forever. Built by one person.</i>
+      <span data-i18n="hero.supportBig">Back the project</span>
+      <i data-i18n="hero.supportSub">GitHub Sponsors</i>
     </a>
   </div>
   <div class="wrap">
     <p class="eyebrow" data-i18n="hero.eyebrow">Free · Open source · Mac</p>
     <h1 data-i18n="hero.h1">Your amps.<br>Your pedals.<br><em>Your laptop.</em></h1>
-    <p class="lede" data-i18n="hero.lede">Every amp you've chased, every cabinet you've mic'd, every pedal you've saved up for — playing back with the feel of the real thing. Plug in and go.</p>
+    <p class="lede" data-i18n="hero.lede">{gear} amps, cabs and pedals running on the laptop you already own. Plug your guitar into your interface and play — free, with no latency you can feel.</p>
     <div class="hero-cta">
       <a class="btn btn-primary btn-lg" href="#download" data-i18n="hero.cta1">Download for Mac</a>
       <a class="btn btn-ghost btn-lg" href="#rig" data-i18n="hero.cta2">See the rig</a>
-      <a class="btn btn-support btn-lg" href="#support" data-i18n="hero.supportBtn">♥ Back OpenRig</a>
+      <a class="btn btn-support btn-lg" href="#support" data-i18n="hero.supportBtn">♥ Back the project</a>
       <span class="hero-note" data-i18n="hero.note">Free forever. No account, no cloud, no subscription.</span>
     </div>
   </div>
   <div class="scroll-hint"><span data-i18n="hero.scroll">Scroll</span><i></i></div>
 </header>
 
-<!-- One file per section under sections/, injected by app.js before the language pass. -->
+<!-- One file per section under sections/, injected by js/sections.js before the language pass. -->
 <div data-include="sections/tone.html"></div>
 <div data-include="sections/stats.html"></div>
 <div data-include="sections/captures.html"></div>
@@ -95,7 +113,7 @@
   </div>
 </footer>
 
-<script src="app.js" defer></script>
+<script type="module" src="js/main.js"></script>
 <!-- Cloudflare Web Analytics — no cookies, no consent banner. -->
 <script defer src="https://static.cloudflareinsights.com/beacon.min.js" data-cf-beacon='{"token": "6b45a12d89e44bfd8e08fe794966ce7a"}'></script>
 </body>
