+++ site/sections/ai.html
@@ -5,7 +5,7 @@
         <p class="eyebrow" data-i18n="ai.eyebrow">First on MCP</p>
         <h2 data-i18n="ai.h2">The first rig<br>an AI can play.</h2>
       </div>
-      <p class="lede rv sub" data-i18n="ai.lede">OpenRig ships its own MCP server — every command an AI needs to build, tweak and drive a rig, exposed the same way for the GUI, a footswitch, or Claude. Point Claude Code, Claude Desktop or Cursor at it and describe the tone you want in plain language.</p>
+      <p class="lede rv sub" data-i18n="ai.lede">OpenRig speaks MCP. Name a song or an artist and Claude looks up which amp, cab and pedals were used, picks the closest match in the catalog and saves the preset onto your rig — while you keep playing.</p>
     </div>
 
     <div class="cards">
