+++ site/sections/captures.html
@@ -3,7 +3,7 @@
     <div class="split" style="align-items:end">
       <div class="stack rv">
         <p class="eyebrow" data-i18n="captures.eyebrow">Captures</p>
-        <h2 data-i18n="captures.h2">A folder of files<br>is not an amp.</h2>
+        <h2 data-i18n="captures.h2">A capture pack<br>becomes one amp.</h2>
       </div>
       <p class="lede rv sub" data-i18n="captures.lede">A capture pack is dozens of files with the gain, the channel and the mic baked into the filename. Everywhere else you audition them one by one. OpenRig's plugins don't ship that list — we read the whole set, work out what actually changes between the files, and hand you those as knobs.</p>
     </div>
