+++ site/sections/gig.html
@@ -10,7 +10,7 @@
     <div class="cards">
       <div class="card rv"><div class="n">01</div><h3 data-i18n="gig.c1.h3">No latency you can feel</h3><p data-i18n="gig.c1.p">The signal goes in and comes back out fast enough that your hands never notice the computer in between.</p></div>
       <div class="card rv"><div class="n">02</div><h3 data-i18n="gig.c2.h3">Silent switching</h3><p data-i18n="gig.c2.p">Swap an amp, add a pedal, change a preset while you're playing. No gaps, no pops, no dead bars.</p></div>
-      <div class="card rv"><div class="n">03</div><h3 data-i18n="gig.c3.h3">No hardware to buy</h3><p data-i18n="gig.c3.p">The laptop you already own is the rig. Nothing extra to buy, nothing extra to carry to the gig.</p></div>
+      <div class="card rv"><div class="n">03</div><h3 data-i18n="gig.c3.h3">No amp head, no pedalboard</h3><p data-i18n="gig.c3.p">You already have an audio interface to plug the guitar into. Everything after it — amp, cab, pedals — runs on the laptop.</p></div>
     </div>
   </div>
 </section>
