+++ site/sections/tone.html
@@ -3,8 +3,8 @@
   <div class="wrap">
     <div class="band-copy rv">
       <div class="kicker-rule"></div>
-      <h2 data-i18n="tone.h2">It has to feel<br>right under<br>your fingers.</h2>
-      <p class="lede" style="margin-top:22px" data-i18n="tone.lede">Pick attack, bloom, the way a note sags when you dig in — that's what makes an amp yours. OpenRig captures it and hands it back, note for note, with no lag between your hands and the speaker.</p>
+      <h2 data-i18n="tone.h2">The amp responds<br>like the real one.</h2>
+      <p class="lede" style="margin-top:22px" data-i18n="tone.lede">Pick attack, sustain, the way a note compresses when you dig in. OpenRig has native amp models written in Rust and it also loads neural captures of real amps — you pick which. Either way there's no latency you can feel between your hand and the speaker.</p>
     </div>
   </div>
 </section>
